% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval and Histogram of Sample Statistics}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations (default is 10000).}

\item{x}{A numeric vector of data to bootstrap from.}

\item{fun}{The function to be applied to each bootstrap sample (e.g., "mean", "median"). Default is "mean".}

\item{alpha}{The significance level for the confidence interval (default is 0.05).}

\item{cx}{Character expansion factor for text annotations (default is 1.5).}

\item{...}{Additional graphical parameters passed to the `hist` function.}
}
\value{
A list with the following elements:
\item{ci}{The bootstrap confidence interval for the specified statistic.}
\item{fun}{The function applied to each bootstrap sample.}
\item{x}{The original data vector.}
}
\description{
Bootstrap Confidence Interval and Histogram of Sample Statistics
}
\examples{
# Bootstrap confidence interval for the mean of a sample
set.seed(123)
sample_data <- rnorm(100, mean = 50, sd = 10)
myboot2(iter = 5000, x = sample_data, fun = "mean", alpha = 0.05, col = "blue")

}
