% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg: Graphical Maximum Likelihood Estimation}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{A log-likelihood function}

\item{theta}{vector of theta values}
}
\value{
The function returns the theta value that maximizes the likelihood and produces a plot
}
\description{
mymaxlikg: Graphical Maximum Likelihood Estimation
}
\examples{
logbin2 <- function(theta) dbinom(4, size = 10, prob = theta, log = TRUE)
theta_vals <- seq(0, 1, by = 0.01)
mymaxlikg(lfun = logbin2, theta = theta_vals)

}
