% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymult.R
\name{mymult}
\alias{mymult}
\title{plot samples from multinomial distribution}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{is the number of times a sample is drawn}

\item{n}{is the size of the sample to be drawn}

\item{p}{is a vector representing the probability of each catagory}
}
\value{
plots residuals vs fitted values
}
\description{
plot samples from multinomial distribution
}
\examples{
mymult(iter = 200, n = 15, p = c(0.5, 0.3, 0.2))

}
