% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{displays area under curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of normal distribution}

\item{a}{value at which probability less than will be calculated}
}
\value{
shaded area, mean, standard deviation, and P(X < a)
}
\description{
displays area under curve
}
\examples{
# Plotting and calculating probability for N(0, 1) where a = 1.5
myncurve(mu = 0, sigma = 1, a = 1.5)

}
