% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{finds optimal number of plane tickets to sell}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of available seats}

\item{gamma}{probability of overbooking}

\item{p}{probability of passenger showing up}
}
\value{
list including nd, nc, N, p, and gamma
}
\description{
finds optimal number of plane tickets to sell
}
