% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resvfit.R
\name{resvfit}
\alias{resvfit}
\title{plot residuals vs fitted values}
\usage{
resvfit(x)
}
\arguments{
\item{x}{a linear model object}
}
\value{
plots residuals vs fitted values
}
\description{
plot residuals vs fitted values
}
