% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{displays area under curve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{value at which probability less than will be calculated}

\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of normal distribution}
}
\value{
shaded area, mean, standard deviation, and P(X < a)
}
\description{
displays area under curve
}
